Adds POST error messages
