+++ src/components/LibraryDetails.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   Box,
   Form,
@@ -10,10 +10,9 @@ import {
 
 import LibraryDetailsGrid from './LibraryDetailsGrid';
 import StageSelect from './StageSelect';
-import { LibraryData } from './RegistryAdmin';
 import useTokenContext from '../context/tokenContext';
 import { UPDATE_LIBRARY_STAGE } from '../constants';
-import useLibrariesContext from '../context/librariesContext';
+import useLibrariesContext, { LibraryData } from '../context/librariesContext';
 
 interface LibraryDetailsProps {
   library: LibraryData;
@@ -29,6 +28,9 @@ const LibraryDetails = ({ library }: LibraryDetailsProps) => {
     areas,
     stages,
   } = library;
+
+  const [libraryStageError, setLibraryStageError] = useState<string>('');
+  const [registryStageError, setRegistryStageError] = useState<string>('');
 
   const { accessToken } = useTokenContext();
   const { setUpdatedLibrary } = useLibrariesContext();
@@ -83,8 +85,8 @@ const LibraryDetails = ({ library }: LibraryDetailsProps) => {
   };
 
   const handleStageChange = (stage: string, value: string) => {
-    // UPDATE_LIBRARY_STAGE endpoint expects the 'Library Stage' and the 'Registry
-    // Stage', even if only one value has been updated.
+    // The UPDATE_LIBRARY_STAGE endpoint expects the 'Library Stage' and the
+    // 'Registry Stage' values, even if only one is being updated.
     const body = new FormData();
     body.append('uuid', uuid);
     body.append(
@@ -96,6 +98,8 @@ const LibraryDetails = ({ library }: LibraryDetailsProps) => {
       stage === 'registryStage' ? value : stages.registry_stage
     );
 
+    const errorMessage = 'Library did not update. Please try again.';
+
     fetch(UPDATE_LIBRARY_STAGE, {
       headers: { Authorization: `Bearer ${accessToken}` },
       method: 'POST',
@@ -103,10 +107,17 @@ const LibraryDetails = ({ library }: LibraryDetailsProps) => {
     })
       .then((response) => {
         if (response.ok) {
+          // Reset errors in state to empty strings just in case they are set
+          // to an old error.
+          setLibraryStageError('');
+          setRegistryStageError('');
           // If the POST request is successful, we'll follow it with a GET request
           // to fetch the library that was just updated.
           fetchUpdatedLibrary();
         } else {
+          stage === 'libraryStage'
+            ? setLibraryStageError(errorMessage)
+            : setRegistryStageError(errorMessage);
           throw new Error('There was an issue updating this library.');
         }
       })
@@ -119,18 +130,20 @@ const LibraryDetails = ({ library }: LibraryDetailsProps) => {
         <FormRow>
           <FormField>
             <StageSelect
-              uuid={uuid}
+              error={libraryStageError}
+              handleStageChange={handleStageChange}
               stage='libraryStage'
               value={stages.library_stage}
-              handleStageChange={handleStageChange}
+              uuid={uuid}
             />
           </FormField>
           <FormField>
             <StageSelect
-              uuid={uuid}
+              error={registryStageError}
+              handleStageChange={handleStageChange}
               stage='registryStage'
               value={stages.registry_stage}
-              handleStageChange={handleStageChange}
+              uuid={uuid}
             />
           </FormField>
         </FormRow>
