+++ src/context/librariesContext.tsx
@@ -1,7 +1,44 @@
-import React, { createContext, Dispatch, useEffect, useState } from 'react';
-import { LibraryData } from '../components/RegistryAdmin';
+import React, {
+  createContext,
+  Dispatch,
+  useContext,
+  useEffect,
+  useState,
+} from 'react';
+export interface LibraryData {
+  areas: {
+    focus: string[];
+    service: string[];
+  };
+  basic_info: {
+    description: string;
+    internal_urn: string;
+    name: string;
+    number_of_patrons: string;
+    online_registration: string;
+    pls_id: string;
+    short_name: string;
+    timestamp: string;
+  };
+  stages: {
+    library_stage: 'testing' | 'production' | 'cancelled';
+    registry_stage: 'testing' | 'production' | 'cancelled';
+  };
+  urls_and_contact: {
+    authentication_url: string;
+    contact_email: string | null;
+    contact_validated: string;
+    copyright_email: string | null;
+    copyright_validated: string;
+    help_email: string | null;
+    help_validated: string;
+    opds_url: string;
+    web_url: string;
+  };
+  uuid: string;
+}
 
-type LibrariesContextValues = {
+export type LibrariesContextValues = {
   libraries: LibraryData[];
   setLibraries: Dispatch<React.SetStateAction<LibraryData[]>>;
   setUpdatedLibrary: Dispatch<React.SetStateAction<LibraryData | undefined>>;
@@ -16,6 +53,7 @@ export const LibrariesContext = createContext<
 >(undefined);
 
 export const LibrariesProvider = ({ children }: LibrariesProviderProps) => {
+  console.log('in libraries provider');
   const [updatedLibrary, setUpdatedLibrary] = useState<
     LibraryData | undefined
   >();
@@ -48,7 +86,7 @@ export const LibrariesProvider = ({ children }: LibrariesProviderProps) => {
 };
 
 const useLibrariesContext = () => {
-  const context = React.useContext(LibrariesContext);
+  const context = useContext(LibrariesContext);
   if (typeof context === 'undefined') {
     throw new Error('useLibraries must be used within a LibraryProvider');
   }
